adding two new functions to convert from ws_obj and gff to genome or metagenome

+++ lib/GenomeFileUtil/core/FastaGFFToGenome.py
@@ -22,6 +22,7 @@ from GenomeFileUtil.core.GenomeInterface import GenomeInterface
 from GenomeFileUtil.core.GenomeUtils import is_parent, warnings, \
     check_full_contig_length_or_multi_strand_feature
 from GenomeFileUtil.core.GenomeUtils import propagate_cds_props_to_gene, load_ontology_mappings
+from GenomeFileUtil.core.MiscUtils import validate_lists_have_same_elements
 from installed_clients.AssemblyUtilClient import AssemblyUtil
 from installed_clients.DataFileUtilClient import DataFileUtil
 
@@ -120,9 +121,12 @@ class FastaGFFToGenome:
 
         # 6) return the result
         info = result['info']
+        prefix = ''
+        if self.is_metagenome:
+            prefix = 'meta'
         details = {
-            'genome_ref': f'{info[6]}/{info[0]}/{info[4]}',
-            'genome_info': info
+            prefix+'genome_ref': f'{info[6]}/{info[0]}/{info[4]}',
+            prefix+'genome_info': info
         }
 
         return details
@@ -160,15 +164,41 @@ class FastaGFFToGenome:
             genome_type = "metagenome"
         else:
             genome_type = params.get('genome_type', 'isolate')
-        assembly_ref = self.au.save_assembly_from_fasta(
-            {'file': {'path': input_fasta_file},
-             'workspace_name': params['workspace_name'],
-             'assembly_name': params['genome_name'] + ".assembly",
-             'type': genome_type,
-             })
-        assembly_data = self.dfu.get_objects(
-            {'object_refs': [assembly_ref],
-             'ignore_errors': 0})['data'][0]['data']
+        # ADD THE OPTION FOR NO ASSEMBLY CREATION HERE
+        if params.get('existing_assembly_ref'):
+            assembly_ref = params['existing_assembly_ref']
+
+            if not re.match("\d+\/\d+\/\d+", assembly_ref):
+                raise ValueError(f"Assembly ref: {assembly_ref} is not a valid format. Must"
+                                 f" be in numerical <ws>/<object>/<version> format.")
+            ret = self.dfu.get_objects(
+                {'object_refs': [assembly_ref]}
+            )['data'][0]
+
+            if "KBaseGenomeAnnotations.Assembly" not in ret['info'][2]:
+                raise ValueError(f"{assembly_ref} is not a reference to an assembly")
+            
+            assembly_data = ret['data']
+            # should do more thorough check of sequences.
+            if not validate_lists_have_same_elements(
+                assembly_data['contigs'].keys(),
+                contig_ids
+            ):
+                raise ValueError(f"provided assembly with ref {assembly_ref} does not "
+                                  "have matching contig ids to provided input fasta.")
+
+            logging.info(f"Using supplied assembly: {assembly_ref}")
+
+        else:
+            assembly_ref = self.au.save_assembly_from_fasta(
+                {'file': {'path': input_fasta_file},
+                 'workspace_name': params['workspace_name'],
+                 'assembly_name': params['genome_name'] + ".assembly",
+                 'type': genome_type,
+                 })
+            assembly_data = self.dfu.get_objects(
+                {'object_refs': [assembly_ref],
+                 'ignore_errors': 0})['data'][0]['data']
 
         # generate genome info
         genome = self._gen_genome_info(assembly_ref, assembly_data,
@@ -262,13 +292,18 @@ class FastaGFFToGenome:
             below seems like weird if statement
             '''
             if file.get('path') is not None:
+                print('freddddyiiie flanstone')
                 local_file_path = file['path']
                 file_path = os.path.join(input_directory, os.path.basename(local_file_path))
                 logging.info(f'Moving file from {local_file_path} to {file_path}')
                 # Metagenome Updates
                 # not sure if we have to be careful about moving the objects
                 # around
-                shutil.copy2(local_file_path, file_path)
+                if os.path.isfile(local_file_path):
+                    shutil.copy2(local_file_path, file_path)
+                else:
+                    raise FileNotFoundError(f"Input {key} file {local_file_path} not found")
+                err_msg  = "Shutil copy unsucessful"
 
             elif file.get('shock_id') is not None:
                 # handle shock file
@@ -279,7 +314,7 @@ class FastaGFFToGenome:
                                                     'shock_id': file['shock_id']
                                                     })['node_file_name']
                 file_path = os.path.join(input_directory, file_name)
-
+                err_msg = "Shock retrieval"
             # extract the file if it is compressed
             '''
             Metagenome Changes:
@@ -289,10 +324,17 @@ class FastaGFFToGenome:
             if file_path is not None:
                 logging.info("staged input file =" + file_path)
                 sys.stdout.flush()
+                if not os.path.isfile(file_path):
+                    raise FileNotFoundError(f"not so fasth buckaroo, {file_path} isnt real")
                 dfUtil_result = self.dfu.unpack_file({'file_path': file_path})
                 file_paths[key] = dfUtil_result['file_path']
+                err_msg = "DataFielUtil 'unpack_file' function call"
             else:
                 raise ValueError('No valid files could be extracted based on the input')
+
+            if not os.path.isfile(file_path):
+                raise ValueError(f"{err_msg} for {key} file to {file_path}")
+
 
         return file_paths
